+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "Playful Sparkle: Dev Toolbox" extension will be documented in this file.
 
+## [0.0.22] - 2025-11-09
+
+- Fixed bug parsing multi-line slug processing.
+- Fixed bug escaping slug separator.
+- Refactored slug generator, betted handling of double extension like `.tar.gz`, support for unicode, LTR/RTL, emoji characters.
+
 ## [0.0.21] - 2025-11-08
 
 - Refactored clean text, the text is now processed by character‑by‑character loop: first normalises malformed Unicode sequences, then uses lookup sets to strip hidden control characters, preserve valid surrogate pairs, and clean up exotic spaces and dashes.